+++ libs/symbols.c
@@ -13,9 +13,9 @@
 #include "string.h"
 
 /* Find the corresponding symbol index in the symbol table according to the address */
-long long symbol_idx_lookup(unsigned long addr)
+long long symbol_idx_lookup(size_t addr)
 {
-    for (unsigned long i = 0; i < sym_count; i++) {
+    for (size_t i = 0; i < sym_count; i++) {
         if (addresses[i] > addr) return i - 1;
     }
     return -1;
@@ -24,7 +24,7 @@ long long symbol_idx_lookup(unsigned long addr)
 /* Find the corresponding symbol index in the symbol table according to the symbol name */
 long long symbol_addr_idx_lookup(char *sym_name)
 {
-    for (unsigned long i = 0; i < sym_count; i++) {
+    for (size_t i = 0; i < sym_count; i++) {
         if (strcmp(symbols[i], sym_name) == 0) return i;
     }
     return -1;
